+++ integration/components/documents/DocumentDetail.tsx
@@ -1,7 +1,5 @@
 import * as React from "react";
-import {isLocalDocument} from "../../DocumentDB";
 import LocalDocument from "./LocalDocument";
-import HostedDocument from "./HostedDocument";
 import {DocumentIDNameResponse} from "../../../ironweb";
 
 interface DocumentDetailProps {
@@ -11,9 +9,6 @@ interface DocumentDetailProps {
 
 export default class DocumentDetail extends React.Component<DocumentDetailProps> {
     render() {
-        if (isLocalDocument(this.props.document.documentID)) {
-            return <LocalDocument document={this.props.document} backToDocumentList={this.props.backToList} />;
-        }
-        return <HostedDocument document={this.props.document} backToDocumentList={this.props.backToList} />;
+        return <LocalDocument document={this.props.document} backToDocumentList={this.props.backToList} />;
     }
 }